+++ app/Http/Controllers/DAV/Backend/CalDAV/CalDAVBirthdays.php
@@ -39,7 +39,7 @@ class CalDAVBirthdays extends AbstractCalDAVBackend
     /**
      * Extension for Calendar objects.
      *
-     * @var string
+     * @return string
      */
     public function getExtension()
     {