+++ resources/lang/cs/validation.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * ⚠️ Edition not allowed except for 'en' language.
+ * ⚠️ Editing not allowed except for 'en' language.
  *
  * @see https://github.com/monicahq/monica/blob/master/docs/contribute/translate.md for translations.
  */
@@ -116,7 +116,7 @@ return [
         'string' => ':attribute musí být přesně :size znaků dlouhý.',
         'array' => ':attribute musí obsahovat právě :size prvků.',
     ],
-    'starts_with' => ':attribute musí začínat jednou z následujících hodnot: :values',
+    'starts_with' => ':attribute musí začínat jednou z následujících hodnot: :values.',
     'string' => ':attribute musí být řetězec znaků.',
     'timezone' => ':attribute musí být platná časová zóna.',
     'unique' => ':attribute musí být unikátní.',