+++ resources/lang/de/dashboard.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * ⚠️ Edition not allowed except for 'en' language.
+ * ⚠️ Editing not allowed except for 'en' language.
  *
  * @see https://github.com/monicahq/monica/blob/master/docs/contribute/translate.md for translations.
  */
@@ -35,7 +35,7 @@ return [
     'statistics_gifts' => 'Geschenke',
 
     'reminders_next_months' => 'Ereignisse der nächsten 3 Monate',
-    'reminders_none' => 'Keine Erinnerungen für diesen Monat',
+    'reminders_none' => 'Keine Erinnerungen für diesen Monat.',
 
     'product_changes' => 'Produktänderungen',
     'product_view_details' => 'Details anzeigen',