+++ resources/lang/it/dashboard.php
@@ -1,14 +1,14 @@
 <?php
 
 /**
- * ⚠️ Edition not allowed except for 'en' language.
+ * ⚠️ Editing not allowed except for 'en' language.
  *
  * @see https://github.com/monicahq/monica/blob/master/docs/contribute/translate.md for translations.
  */
 
 return [
     'dashboard_blank_title' => 'Benvenuto nel tuo account!',
-    'dashboard_blank_description' => 'Con Monica puoi organizzare tutte le interazioni con le persone a cui tieni.',
+    'dashboard_blank_description' => 'Monica è il luogo per organizzare tutte le interazioni che hai con le persone a cui tieni.',
     'dashboard_blank_cta' => 'Aggiungi il tuo primo contatto',
     'dashboard_blank_illustration' => 'Illustrazione by <a href=":url">Freepik</a>',
 
@@ -16,11 +16,11 @@ return [
 
     'tab_recent_calls' => 'Chiamate recenti',
     'tab_favorite_notes' => 'Note preferite',
-    'tab_calls_blank' => 'Non hai ancora aggiunto nessuna etichetta.',
+    'tab_calls_blank' => 'Non hai ancora registrato alcuna chiamata.',
     'tab_debts' => 'Debiti',
-    'tab_debts_blank' => 'Non hai inserito ancora nessun debito.',
+    'tab_debts_blank' => 'Non hai ancora registrato alcun debito.',
     'tab_tasks' => 'Promemoria',
-    'tab_tasks_blank' => 'Non hai registrato ancora nessuna attività.',
+    'tab_tasks_blank' => 'Non hai ancora alcuna attività.',
 
     'tasks_add_task_placeholder' => 'Cosa vuoi ricordarti?',
     'tasks_tab_your_contacts' => 'Promemoria riguardanti i tuoi contatti',
@@ -35,7 +35,7 @@ return [
     'statistics_gifts' => 'Regali',
 
     'reminders_next_months' => 'Eventi nei prossimi 3 mesi',
-    'reminders_none' => 'Nessun promemoria per questo mese',
+    'reminders_none' => 'Nessun promemoria per questo mese.',
 
     'product_changes' => 'Changelog',
     'product_view_details' => 'Mostra dettagli',