+++ resources/lang/tr/dashboard.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * ⚠️ Edition not allowed except for 'en' language.
+ * ⚠️ Editing not allowed except for 'en' language.
  *
  * @see https://github.com/monicahq/monica/blob/master/docs/contribute/translate.md for translations.
  */
@@ -35,7 +35,7 @@ return [
     'statistics_gifts' => 'Hediyeler',
 
     'reminders_next_months' => 'Gelecek 3 aydaki etkinlikler',
-    'reminders_none' => 'Bu ay için hatırlatma yok',
+    'reminders_none' => 'Bu ay için hatırlatma yok.',
 
     'product_changes' => 'Ürün değişiklikleri',
     'product_view_details' => 'Detayları göster',