+++ resources/lang/vi/dashboard.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * ⚠️ Edition not allowed except for 'en' language.
+ * ⚠️ Editing not allowed except for 'en' language.
  *
  * @see https://github.com/monicahq/monica/blob/master/docs/contribute/translate.md for translations.
  */
@@ -20,7 +20,7 @@ return [
     'tab_debts' => 'Khoản nợ',
     'tab_debts_blank' => 'Bạn chưa ghi khoản nợ nào.',
     'tab_tasks' => 'Tasks',
-    'tab_tasks_blank' => 'You haven’t any task yet.',
+    'tab_tasks_blank' => 'You haven’t any tasks yet.',
 
     'tasks_add_task_placeholder' => 'What is this task about?',
     'tasks_tab_your_contacts' => 'Tasks related to your contacts',
@@ -35,7 +35,7 @@ return [
     'statistics_gifts' => 'Quà tặng',
 
     'reminders_next_months' => 'Sự kiện trong 3 tháng tới',
-    'reminders_none' => 'Không có nhắc nhở nào trong tháng này',
+    'reminders_none' => 'Không có nhắc nhở nào trong tháng này.',
 
     'product_changes' => 'Cập nhật sản phẩm',
     'product_view_details' => 'Xem chi tiết',