+++ resources/views/people/food-preferences/index.blade.php
@@ -1,27 +1,18 @@
 <div class="sidebar-box">
 
+  <div class="sidebar-box-title">
+    <h3>{{ trans('people.food_preferences_title') }}</h3>
+  </div>
+
+  <p class="sidebar-box-paragraph">
+
   @if (is_null($contact->food_preferences))
-
-    <p class="sidebar-box-title">
-      <strong>{{ trans('people.food_preferences_title') }}</strong>
-    </p>
-
-    <p class="sidebar-box-paragraph">
-      <a href="{{ route('people.food.index', $contact) }}">{{ trans('app.add') }}</a>
-    </p>
-
+    <a href="{{ route('people.food.index', $contact) }}">{{ trans('app.add') }}</a>
   @else
-
-    <p class="sidebar-box-title">
-      <strong>{{ trans('people.food_preferences_title') }}</strong>
-    </p>
-
-    {{-- Information about the significant other --}}
-    <p class="sidebar-box-paragraph">
-      {{ $contact->food_preferences }}
-      <a href="{{ route('people.food.index', $contact) }}" class="action-link">{{ trans('app.edit') }}</a>
-    </p>
-
+    {{ $contact->food_preferences }}
+    <a href="{{ route('people.food.index', $contact) }}" class="action-link">{{ trans('app.edit') }}</a>
   @endif
+
+  </p>
 
 </div>