+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ### New features:
 
+* Add an artisan command to create an account
 * Add the notion of AddressBooks
 * Allow customization of life event types
 
 ### Enhancements:
 
 * Use Unknown as default gender
+* Better padding and text about Me as a contact on Settings page
 
 ### Fixes: